moved float functions to float package
