+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/FloatMatchers.kt
@@ -2,34 +2,8 @@ package io.kotlintest.matchers
 
 import io.kotlintest.Matcher
 import io.kotlintest.Result
-import io.kotlintest.shouldBe
-import io.kotlintest.shouldNotBe
 
 infix fun Float.plusOrMinus(tolerance: Float): FloatToleranceMatcher = FloatToleranceMatcher(this, tolerance)
-
-fun exactly(d: Float): Matcher<Float> = object : Matcher<Float> {
-  override fun test(value: Float) = Result(value == d, "$value is not equal to expected value $d", "$value should not be equal to $d")
-}
-
-fun lt(x: Float) = beLessThan(x)
-fun beLessThan(x: Float) = object : Matcher<Float> {
-  override fun test(value: Float) = Result(value < x, "$value should be < $x", "$value should not be < $x")
-}
-
-fun lte(x: Float) = beLessThanOrEqualTo(x)
-fun beLessThanOrEqualTo(x: Float) = object : Matcher<Float> {
-  override fun test(value: Float) = Result(value <= x, "$value should be <= $x", "$value should not be <= $x")
-}
-
-fun gt(x: Float) = beGreaterThan(x)
-fun beGreaterThan(x: Float) = object : Matcher<Float> {
-  override fun test(value: Float) = Result(value > x, "$value should be > $x", "$value should not be > $x")
-}
-
-fun gte(x: Float) = beGreaterThanOrEqualTo(x)
-fun beGreaterThanOrEqualTo(x: Float) = object : Matcher<Float> {
-  override fun test(value: Float) = Result(value >= x, "$value should be >= $x", "$value should not be >= $x")
-}
 
 class FloatToleranceMatcher(val expected: Float, val tolerance: Float) : Matcher<Float> {
 
@@ -51,6 +25,3 @@ class FloatToleranceMatcher(val expected: Float, val tolerance: Float) : Matcher
 
   infix fun plusOrMinus(tolerance: Float): FloatToleranceMatcher = FloatToleranceMatcher(expected, tolerance)
 }
-
-infix fun Float.shouldBeExactly(x: Float) = this shouldBe exactly(x)
-infix fun Float.shouldNotBeExactly(x: Float) = this shouldNotBe exactly(x)
