+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/floats/matchers.kt
@@ -1,12 +1,37 @@
 package io.kotlintest.matchers.floats
 
-import io.kotlintest.matchers.exactly
+import io.kotlintest.Matcher
+import io.kotlintest.Result
 import io.kotlintest.matchers.gt
 import io.kotlintest.matchers.gte
 import io.kotlintest.matchers.lt
 import io.kotlintest.matchers.lte
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldNotBe
+
+fun exactly(d: Float): Matcher<Float> = object : Matcher<Float> {
+  override fun test(value: Float) = Result(value == d, "$value is not equal to expected value $d", "$value should not be equal to $d")
+}
+
+fun lt(x: Float) = beLessThan(x)
+fun beLessThan(x: Float) = object : Matcher<Float> {
+  override fun test(value: Float) = Result(value < x, "$value should be < $x", "$value should not be < $x")
+}
+
+fun lte(x: Float) = beLessThanOrEqualTo(x)
+fun beLessThanOrEqualTo(x: Float) = object : Matcher<Float> {
+  override fun test(value: Float) = Result(value <= x, "$value should be <= $x", "$value should not be <= $x")
+}
+
+fun gt(x: Float) = beGreaterThan(x)
+fun beGreaterThan(x: Float) = object : Matcher<Float> {
+  override fun test(value: Float) = Result(value > x, "$value should be > $x", "$value should not be > $x")
+}
+
+fun gte(x: Float) = beGreaterThanOrEqualTo(x)
+fun beGreaterThanOrEqualTo(x: Float) = object : Matcher<Float> {
+  override fun test(value: Float) = Result(value >= x, "$value should be >= $x", "$value should not be >= $x")
+}
 
 infix fun Float.shouldBeLessThan(x: Float) = this shouldBe lt(x)
 infix fun Float.shouldNotBeLessThan(x: Float) = this shouldNotBe lt(x)
