+++ kotlintest-tests/kotlintest-tests-core/src/test/kotlin/com/sksamuel/kotlintest/matchers/floats/FloatMatchersTest.kt
@@ -1,6 +1,5 @@
 package com.sksamuel.kotlintest.matchers.floats
 
-import io.kotlintest.matchers.doubles.shouldNotBeLessThanOrEqual
 import io.kotlintest.matchers.floats.*
 import io.kotlintest.specs.StringSpec
